server: Change some fuel variables to double type.

+++ src/server/netserver.cpp
@@ -1289,8 +1289,8 @@ int Send_self(connection_t *connp,
                       pl->check,
 
                       pl->fuel.current,
-                      pl->fuel.sum >> FUEL_SCALE_BITS,
-                      pl->fuel.max >> FUEL_SCALE_BITS,
+                      ((long)pl->fuel.sum) >> FUEL_SCALE_BITS,
+                      ((long)pl->fuel.max) >> FUEL_SCALE_BITS,
 
                       connp->view_width, connp->view_height,
                       connp->debris_colors,
